Use buffered io for performance
In the previous implementation we'd have to read all of the contents of
the buffer into memory before we start printing. If the given number of
bytes was too big, that also consumed to much memory.

With buffered IO we limit the memory consumption and we get to start
printing right away, making it faster.